+++ src/wrapt/_wrappers.c
@@ -1224,6 +1224,17 @@ static PyObject *WraptObjectProxy_deepcopy(
 
 /* ------------------------------------------------------------------------- */
 
+static PyObject *WraptObjectProxy_reduce(
+        WraptObjectProxyObject *self, PyObject *args, PyObject *kwds)
+{
+    PyErr_SetString(PyExc_NotImplementedError,
+                    "object proxy must define __reduce__() or __reduce_ex__()");
+
+    return NULL;
+}
+
+/* ------------------------------------------------------------------------- */
+
 static PyObject *WraptObjectProxy_bytes(
         WraptObjectProxyObject *self, PyObject *args)
 {
@@ -1712,6 +1723,8 @@ static PyMethodDef WraptObjectProxy_methods[] = {
                     METH_NOARGS, 0 },
     { "__deepcopy__", (PyCFunction)WraptObjectProxy_deepcopy,
                     METH_VARARGS | METH_KEYWORDS, 0 },
+    { "__reduce__", (PyCFunction)WraptObjectProxy_reduce,
+                    METH_NOARGS, 0 },
     { "__getattr__", (PyCFunction)WraptObjectProxy_getattr,
                     METH_VARARGS , 0 },
     { "__bytes__",  (PyCFunction)WraptObjectProxy_bytes, METH_NOARGS, 0 },